rewrite the config storage system
Removed the use of the qml localStorage library that was storing the
config as json in a sqlite database in a location with an ugly path.

The configuration is now stored in `~/.config/cool-retro-term/` in json
files.

the `profiles` sub-directory can also be used to add *readonly* profiles
that will be loaded as "builtin".